# [MRG] Numpy pickle to single file