--- a/continuous_integration/install.sh
+++ b/continuous_integration/install.sh
@@ -73,6 +73,11 @@ if [ -z "$NUMPY_VERSION" ]; then
     # setup.cfg so doing it the hacky way ...
     cat setup.cfg | grep -v 'with-doctest=' > setup.cfg.new
     mv setup.cfg{.new,}
+else
+    # We want to ensure no memory copies are performed only when numpy is
+    # installed. This also ensures that we don't keep a strong dependency on
+    # memory_profiler.
+    pip install memory_profiler
 fi
 
 if [[ "$COVERAGE" == "true" ]]; then