--- a/joblib/__init__.py
+++ b/joblib/__init__.py
@@ -115,7 +115,7 @@ Main features
 # Dev branch marker is: 'X.Y.dev' or 'X.Y.devN' where N is an integer.
 # 'X.Y.dev0' is the canonical version of 'X.Y.dev'
 #
-__version__ = '0.9.4'
+__version__ = '0.10.0.dev0'
 
 
 from .memory import Memory, MemorizedResult