--- a/joblib/test/common.py
+++ b/joblib/test/common.py
@@ -7,6 +7,7 @@ import nose
 import time
 import os
 import sys
+import gc
 
 from joblib._multiprocessing_helpers import mp
 from nose import SkipTest
@@ -17,19 +18,40 @@ try:
     import numpy as np
 
     def with_numpy(func):
-        """ A decorator to skip tests requiring numpy.
-        """
+        """A decorator to skip tests requiring numpy."""
         return func
 
 except ImportError:
     def with_numpy(func):
-        """ A decorator to skip tests requiring numpy.
-        """
+        """A decorator to skip tests requiring numpy."""
         def my_func():
             raise nose.SkipTest('Test requires numpy')
         return my_func
     np = None
 
+
+# we use memory_profiler library for memory consumption checks
+try:
+    from memory_profiler import memory_usage
+
+    def with_memory_profiler(func):
+        """A decorator to skip tests requiring memory_profiler."""
+        return func
+
+    def memory_used(func, *args, **kwargs):
+        """Compute memory usage when executing func."""
+        gc.collect()
+        mem_use = memory_usage((func, args, kwargs), interval=.001)
+        return max(mem_use) - min(mem_use)
+
+except ImportError:
+    def with_memory_profiler(func):
+        """A decorator to skip tests requiring memory_profiler."""
+        def dummy_func():
+            raise nose.SkipTest('Test requires memory_profiler.')
+        return dummy_func
+
+    memory_usage = memory_used = None
 
 # A utility to kill the test runner in case a multiprocessing assumption
 # triggers an infinite wait on a pipe by the master process for one of its