--- a/joblib/test/test_pool.py
+++ b/joblib/test/test_pool.py
@@ -328,7 +328,7 @@ def test_memmaping_pool_for_large_arrays():
         # without per-child memory copies
         assert_true(os.path.isdir(p._temp_folder))
         dumped_filenames = os.listdir(p._temp_folder)
-        assert_equal(len(dumped_filenames), 2)
+        assert_equal(len(dumped_filenames), 1)
 
         # Check that memory mapping is not triggered for arrays with
         # dtype='object'
@@ -388,8 +388,8 @@ def test_memmaping_on_dev_shm():
         assert_equal(a.nbytes, 800)
         p.map(id, [a] * 10)
         # a should have been memmaped to the pool temp folder: the joblib
-        # pickling procedure generate a .pkl and a .npy file:
-        assert_equal(len(os.listdir(pool_temp_folder)), 2)
+        # pickling procedure generate one .pkl file:
+        assert_equal(len(os.listdir(pool_temp_folder)), 1)
 
         # create a new array with content that is different from 'a' so that
         # it is mapped to a different file in the temporary folder of the
@@ -398,7 +398,7 @@ def test_memmaping_on_dev_shm():
         assert_equal(b.nbytes, 800)
         p.map(id, [b] * 10)
         # A copy of both a and b are now stored in the shared memory folder
-        assert_equal(len(os.listdir(pool_temp_folder)), 4)
+        assert_equal(len(os.listdir(pool_temp_folder)), 2)
 
     finally:
         # Cleanup open file descriptors