--- a/setup.cfg
+++ b/setup.cfg
@@ -30,3 +30,6 @@ universal=1
 # http://pep8.readthedocs.org/en/latest/intro.html#error-codes
 # E402: module level import not at top of file
 ignore=E402
+# __init__.py is full of "imported but unused" F401 errors which we don't care
+# we exclude this file on purpose.
+exclude = joblib/__init__.py