… CHANGES.rst
… benchmarks/bench_pickle.py
… doc/persistence.rst
… joblib/numpy_pickle.py
… joblib/numpy_pickle_utils.py
… joblib/test/data/joblib_0.10.0_compressed_pickle_py26_np16.gz
… joblib/test/data/joblib_0.10.0_compressed_pickle_py27_np17.gz
… joblib/test/data/joblib_0.10.0_compressed_pickle_py33_np18.gz
… joblib/test/data/joblib_0.10.0_compressed_pickle_py34_np19.gz
… joblib/test/data/joblib_0.10.0_compressed_pickle_py35_np19.gz
… joblib/test/data/joblib_0.10.0_pickle_py26_np16.pkl
… joblib/test/data/joblib_0.10.0_pickle_py26_np16.pkl.bz2
… joblib/test/data/joblib_0.10.0_pickle_py26_np16.pkl.gzip
… joblib/test/data/joblib_0.10.0_pickle_py26_np16.pkl.lzma
… joblib/test/data/joblib_0.10.0_pickle_py26_np16.pkl.xz
… joblib/test/data/joblib_0.10.0_pickle_py27_np17.pkl
… joblib/test/data/joblib_0.10.0_pickle_py27_np17.pkl.bz2
… joblib/test/data/joblib_0.10.0_pickle_py27_np17.pkl.gzip
… joblib/test/data/joblib_0.10.0_pickle_py27_np17.pkl.lzma
… joblib/test/data/joblib_0.10.0_pickle_py27_np17.pkl.xz
… joblib/test/data/joblib_0.10.0_pickle_py33_np18.pkl
… joblib/test/data/joblib_0.10.0_pickle_py33_np18.pkl.bz2
… joblib/test/data/joblib_0.10.0_pickle_py33_np18.pkl.gzip
… joblib/test/data/joblib_0.10.0_pickle_py33_np18.pkl.lzma
… joblib/test/data/joblib_0.10.0_pickle_py33_np18.pkl.xz
… joblib/test/data/joblib_0.10.0_pickle_py34_np19.pkl
… joblib/test/data/joblib_0.10.0_pickle_py34_np19.pkl.bz2
… joblib/test/data/joblib_0.10.0_pickle_py34_np19.pkl.gzip
… joblib/test/data/joblib_0.10.0_pickle_py34_np19.pkl.lzma
… joblib/test/data/joblib_0.10.0_pickle_py34_np19.pkl.xz