@GaelVaroquaux 2015-10-24
> the read/write is slightly slower (31s dump, 5.5s read).

How does it compare to what we currently have?