@aabadie 2015-10-26
> How does it compare to what we currently have?
It adds an overhead of read/write time of approximately 10% : 28s write with the actual made against 31s with this PR  (4.8s against 5s for reading).
Note that this PR is not fully complete, it misses improvements:
* the control file (containing the main serialized object without arrays/matrices) is not compressed
* the whole object (including arrays) could be serialized in a single file using a smart seeking in the file. This could be part of an extra PR.