@aabadie 2015-11-12
> the control file (containing the main serialized object without arrays/matrices) is not compressed

This was done in 4f58113

> the whole object (including arrays) could be serialized in a single file using a smart seeking in the file. This could be part of an extra PR.

I just pushed caeb704 which does this. But it has several drawbacks:
* tests are broken because the actual state doesn't support memory mapping
* I had to patch numpy locally in order to make the gzip file compression work. Here, numpy overload `GzipFile` object `seek()` function in order to be able to seek using negative values. The problem is that numpy assumes the file only contains one array which is not our case here. The only solution is to include numpy array load/save implementation in joblib.

> last, but not least : this PR should also contains a retrocompatibility mecanism with previous versions of the cache format.

This is done in this branch (works in 4f58113)