@GaelVaroquaux 2015-11-12
>     file only contains one array which is not our case here. The only
>     solution is to include numpy array load/save implementation in
>     joblib.

Can we then only reimplement the minimal amount that we need, and not
copy the whole codebase?