@aabadie 2015-11-12
> Can we then only reimplement the minimal amount that we need, and not
copy the whole codebase?

Yes, that's the plan. It should also helps with the memmap case.