@aabadie 2015-11-13
As the status of this PR is imho in a good shape, I changed the status to MRG and renamed to better reflect what it contains : now object are serialized in a single file, even compressed file and memory maps.
Note that there is a basic retrocompatibility mecanism which means that the future new version of joblib will be able to read files containing object cached with older versions. The dump will use the new format version.

All remaining problems mentionned above are solved:
* as suggested, the gzip.GzipFile issue was solved by reimplementing (not to say copy-pasting), the minimum amount of required code from numpy in joblib. I'm not if I did that the best way, but at least it works now. The compression level is 3 (it can be from 0 to 9 as said in the doc).
* the memory mapping works as well using the same technique. I just had to slightly change the way the implementation of the numpy `open_memmap` function : I introduced a parameter for setting the offset in the file where the numpy array is serialized.

Maybe some more tests could be added. I'll post here some memory/speed benchmarks just to compare with the actual implmentation.

Appveyor is stuck, I don't know why.

Waiting for you comments.