@GaelVaroquaux 2015-11-13
For benchmarks, have a look at the following link (hint: there is a link
to a gist with code to benchmark in it)
http://gael-varoquaux.info/programming/joblib-beta-release-fast-compressed-persistence-python-3.html