@aabadie 2015-11-17
As promised to @GaelVaroquaux, I tested gist locally, just to compare with other implementations.

See the results:
* Read speed with this PR:
![compare_libraries_read](https://cloud.githubusercontent.com/assets/1375137/11215600/76646644-8d47-11e5-9c7d-1e5f6e6c2a78.png)
* Read speed with master:
![compare_libraries_read_master](https://cloud.githubusercontent.com/assets/1375137/11215616/84c727c6-8d47-11e5-848c-e4345e09c5a3.png)
* Write speed with this PR:
![compare_libraries_write](https://cloud.githubusercontent.com/assets/1375137/11215645/9abe46ae-8d47-11e5-9c9e-783753d8891c.png)
* write speed with master:
![compare_libraries_write_master](https://cloud.githubusercontent.com/assets/1375137/11215650/a008edf8-8d47-11e5-8809-45b0491d3aad.png)

The conclusion is that the speed is clearly slower with this PR. And as you noticed in your blog post, gzip.GZipFile might not be an option for cache compression.