@aabadie 2015-11-17
> Looking at the plots, it seems that the datasets labels are mixed. I'll update them.

Replying to myself. It double checked and all is ok. It's just the write speed using F-Contiguous data which has a significant impact on performance (especially when uwing pytables with zlib).