@GaelVaroquaux 2015-11-17
It doesn't seem to me that the impact of this PR on read/write speed is very large.