@GaelVaroquaux 2015-11-17
As a side note, when this PR is merged, it would be useful to test lzma for compression under Python 3: it could be much faster. I have added an enhancement issue for this: #273.