@aabadie 2015-11-17
I also want to try the mmap case using read-write mode and especially when the persisted object contains multiple arrays. I'm wondering if the arrays could overlap at the end of the file resulting in a corrupted file.