@GaelVaroquaux 2015-11-17
> I'm wondering if the arrays could overlap at the end of the file
> resulting in a corrupted file.

Sound like we need a test :)