@lesteve 2015-11-17
Out of interest have you checked the memory usage of this PR? I am suspecting that everything will be fine since we are using np.save and np.load but it'd be good to make sure this is the case.