@lesteve 2015-11-17
> As promised to @GaelVaroquaux, I tested your gist locally, just to compare with other implementations.

> See the results:

I would compare master against this PR on the same plot use a single dataset (either MNI or Juelich), and possibly removing all the pytables stuff to make the comparison easier.

According to the original blog post the arrays are of the order of ~100MB. Maybe we should try with bigger arrays and/or multiple arrays.