@lesteve 2015-11-17
You'll need to add generated pickles in joblib/test/data. The script joblib/test/data/create_numpy_pickle.py should help you. It would be great to use similar numpy versions as the one that were used for the joblib 0.9.2 ones.

By the way we probably want to bump up the version to 0.10.0 because of this change of numpy pickle format.