@GaelVaroquaux 2015-11-17
> By the way we probably want to bump up the version to 0.10.0 because of this
> change of numpy pickle format.

Indeed. And that why I would like to squeeze in the use of lzma before
the release, to avoid having multiple changes of formats.