@aabadie 2015-11-18
> Out of interest have you checked the memory usage of this PR?

Yes, the memory footprint is stable.