@aabadie 2015-11-19
> Sound like we need a test :)

@GaelVaroquaux, I just added a test that raises an issue when unpickling an object containing multiple arrays using memory map => the next array offset was not correctly updated and the returned array could be wrong.