@aabadie 2015-11-19
> I would compare master against this PR on the same plot use a single dataset (either MNI or Juelich), and possibly removing all the pytables stuff to make the comparison easier.

Here are some comparison with master. When using the compression, this PR is clearly slower than the actual implementation in master.

![compare_impl_read](https://cloud.githubusercontent.com/assets/1375137/11277741/0f6812ae-8ee8-11e5-996f-53e4f4adb0e4.png)
![compare_impl_write](https://cloud.githubusercontent.com/assets/1375137/11277742/0f69633e-8ee8-11e5-8cfe-e636afa5a189.png)