@GaelVaroquaux 2015-11-19
It's particularly slower in read. It would be interesting to understand why. I don't see a fundemental reason for this to happen (other than the fact that master has been carefully crafted).