@lesteve 2015-11-19
> Here are some comparison with master

zlib3 is the default (i.e. compressed=True). I think you should show it in your comparison plot, maybe show zlib3 rather than zlib1.