@aabadie 2015-11-20
> Also what is the unit of the y axis?

seconds