@aabadie 2015-12-16
> Looking good, can you post your results master vs PR from your benchmark script in a nicely formatted table as a summary?

Here it is:
To compare the actual master and this PR, I ran several benches that measure read and write speeds, memory consumption with different kind of data (sample nifti fsl data, big array, dictionnary, etc), with/without compression and when using memory map.

* Read/write time comparison with sample fsl data:
  * Read: results are the same between master and this PR.
![master_nifti_read](https://cloud.githubusercontent.com/assets/1375137/11838892/2aa64a10-a3ea-11e5-9a97-1ac6f0e67b7c.png)

  * Write: this PR is slightly faster.
![master_nifti_write](https://cloud.githubusercontent.com/assets/1375137/11838907/3b826670-a3ea-11e5-9ee4-1c95caf03958.png)

* Read/write time comparison with generated arrays/dictionary:  
Big array size is 190.7 MiB, big dictionary has 10000 keys with random values, 2 big arrays takes 367MiB and the combo combines 2 big arrays and the big dictionary: 
  * Read: results are slightly faster with PR but not with the cases where there's the big dictionary. I'm not sure where this comes from. The only difference is the way we compressed data is unpickled : in master the full compressed pickle is read in a row whereas in this PR we uncompress as pickle access the file (through the GzipFile object). So my intuition if that pickle needs to access several times the file object when unpickling the dictionary values and that could explain why this is slower.
![master_random_read](https://cloud.githubusercontent.com/assets/1375137/11839053/38bf1e5a-a3eb-11e5-8658-8bd1d8c95472.png)
  * Write: results are faster with this PR except in the cases with big dictionaries (probably because of the same reason I mentioned above).
![master_random_write](https://cloud.githubusercontent.com/assets/1375137/11839079/59bbd044-a3eb-11e5-9af0-2f4695ee4812.png)

In term of __memory consumption__, the main difference comes the compress case. With memory map and non compressed pickle, both implementation have no memory copy. Here are the results I computed using `memory_profiler` with 2 types of datasets: big array (190.7MiB) and big dictionary (10000 keys) and with compressed/non compressed pickle:

* Big array:

  * master - non compressed 
![master_array](https://cloud.githubusercontent.com/assets/1375137/11840340/62861fae-a3f5-11e5-934d-e11cb2b4d51e.png) 
  * pr 260 non compressed
![pr260_array](https://cloud.githubusercontent.com/assets/1375137/11840347/67bf1fe8-a3f5-11e5-81c9-3ce330003995.png)

  * master - compressed
![master_array_compress](https://cloud.githubusercontent.com/assets/1375137/11840992/657dca68-a3fa-11e5-8d6b-b773b969ffae.png) 
  * pr 260 compressed 
![pr260_array_compress](https://cloud.githubusercontent.com/assets/1375137/11840996/6b301d62-a3fa-11e5-95c7-7fec08831797.png)

The results are the same using non compressed pickles but with compressed pickle this PR is clearly better. This is because array are now written/read by chunks.

* Big dict:

 * master - non compressed
![master_dict](https://cloud.githubusercontent.com/assets/1375137/11841071/df8fc2b6-a3fa-11e5-8803-82fa978439fa.png)
 * pr 260 non compressed
![pr260_dict](https://cloud.githubusercontent.com/assets/1375137/11841072/e50b44a4-a3fa-11e5-968b-0ac44c48e4ab.png)
 * master - compressed
![master_dict_compress](https://cloud.githubusercontent.com/assets/1375137/11841075/eaa2e0d4-a3fa-11e5-9e7a-ea7d0e119156.png)
  * pr 260 compressed
![pr260_dict_compress](https://cloud.githubusercontent.com/assets/1375137/11841079/ef5e45aa-a3fa-11e5-97d0-5f54142bf2e3.png)

With the big dictionary, one clearly sees that there are memory copies but they comes from the pickle implementation and we don't control this.
In the compressed case, the write time increase coming from this PR is clearly visible.



Here are the full value tables used to generate the charts and plots:

* With master:

Dataset | strategy | write time | read time |  write max memory | read max memory | disk
--------- | ----------- | ------ | ------ | -------------- | -------------- | -----
MNI(C) | compress0 | 0.069 | 0.1378 | 0.1 | 62.0 | 62.0
MNI(C) | compress3 | 0.507 | 0.1398 | 66.3 | 62.0 | 3.7
MNI(C) | mmapr | 0.083 | 0.0042 | 0.0 | 0.0 | 62.0
MNI(F) | compress0 | 0.086 | 0.1399 | 0.0 | 62.0 | 62.0
MNI(F) | compress3 | 0.323 | 0.1266 | 64.3 | 61.9 | 2.1
MNI(F) | mmapr | 0.101 | 0.0038 | 0.0 | 0.0 | 62.0
Juelich(C) | compress0 | 0.217 | 0.2304 | 0.0 | 104.2 | 104.2
Juelich(C) | compress3 | 0.489 | 0.1511 | 107.1 | 104.2 | 1.6
Juelich(C) | mmapr | 0.199 | 0.0051 | 0.0 | 0.0 | 104.2
Juelich(F) | compress0 | 0.186 | 0.2325 | 0.0 | 104.2 | 104.2
Juelich(F) | compress3 | 0.412 | 0.1462 | 107.1 | 104.2 | 0.9
Juelich(F) | mmapr | 0.212 | 0.0051 | 0.0 | 0.0 | 104.2
Bigarray | compress0 | 0.440 | 0.4195 | 0.2 | 190.7 | 190.7
Bigarray | compress3 | 7.955 | 1.5638 | 549.5 | 371.6 | 180.9
Bigarray | mmapr | 0.442 | 0.0022 | 0.0 | 0.0 | 190.7
2bigarrays | compress0 | 0.572 | 0.5915 | 0.0 | 267.0 | 267.0
2bigarrays | compress3 | 11.673 | 2.2609 | 552.7 | 371.5 | 253.2
2bigarrays | mmapr | 0.545 | 0.0044 | 0.0 | 0.0 | 267.0
Bigdict | compress0 | 0.454 | 0.2079 | 5.8 | 0.0 | 0.4
Bigdict | compress3 | 0.470 | 0.2174 | 5.4 | 0.0 | 0.2
Bigdict | mmapr | 0.488 | 0.1982 | 4.5 | 0.4 | 0.4
Dict/arrays | compress0 | 0.971 | 0.8413 | 4.2 | 267.0 | 267.5
Dict/arrays | compress3 | 12.001 | 2.3410 | 550.9 | 371.9 | 253.5
Dict/arrays | mmapr | 0.939 | 0.2286 | 3.9 | 0.0 | 267.5

* With this PR:

Dataset | strategy | write time | read time |  write max memory | read max memory | disk
--------- | ----------- | ------ | ------ | -------------- | -------------- | -----
MNI(C) | compress0 | 0.081 | 0.1370 | 0.2 | 62.0 | 62.0
MNI(C) | compress3 | 0.498 | 0.1401 | 0.0 | 61.9 | 3.7
MNI(C) | mmapr | 0.061 | 0.0025 | 0.0 | 0.0 | 62.0
MNI(F) | compress0 | 0.051 | 0.1345 | 0.0 | 62.0 | 62.0
MNI(F) | compress3 | 0.280 | 0.1227 | 0.0 | 61.9 | 2.1
MNI(F) | mmapr | 0.055 | 0.0019 | 0.0 | 0.0 | 62.0
Juelich(C) | compress0 | 0.156 | 0.2345 | 0.0 | 104.2 | 104.2
Juelich(C) | compress3 | 0.417 | 0.1622 | 0.0 | 104.1 | 1.6
Juelich(C) | mmapr | 0.128 | 0.0019 | 0.0 | 0.0 | 104.2
Juelich(F) | compress0 | 0.164 | 0.2271 | 0.0 | 104.2 | 104.2
Juelich(F) | compress3 | 0.363 | 0.1486 | 0.0 | 104.1 | 0.9
Juelich(F) | mmapr | 0.148 | 0.0016 | 0.0 | 0.0 | 104.2
Bigarray | compress0 | 0.359 | 0.4174 | 0.1 | 190.7 | 190.7
Bigarray | compress3 | 7.741 | 1.4000 | 33.2 | 190.6 | 180.9
Bigarray | mmapr | 0.337 | 0.0016 | 0.0 | 0.0 | 190.7
2bigarrays | compress0 | 2.025 | 0.5886 | 0.0 | 267.0 | 267.0
2bigarrays | compress3 | 11.088 | 1.9262 | 33.0 | 267.0 | 253.2
2bigarrays | mmapr | 0.521 | 0.0017 | 0.0 | 0.0 | 267.0
Bigdict | compress0 | 0.466 | 0.2204 | 5.7 | 0.0 | 0.4
Bigdict | compress3 | 0.926 | 0.5403 | 5.0 | 0.0 | 0.2
Bigdict | mmapr | 0.478 | 0.1960 | 4.5 | 0.0 | 0.4
Dict/arrays | compress0 | 1.300 | 0.8519 | 4.5 | 267.0 | 267.5
Dict/arrays | compress3 | 11.859 | 2.4806 | 37.9 | 267.0 | 253.4
Dict/arrays | mmapr | 0.993 | 0.2161 | 4.0 | 0.0 | 267.5